Remove workarounds facilitating support of legacy nuget.commandline
